requirements.txt
astroid==2.7.2
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
conllu==4.4
cycler==0.10.0
cymem==2.0.5
empath==0.89
en-core-sci-lg==0.4.0
en-core-sci-scibert==0.4.0
en-core-sci-sm==0.4.0
en-ner-bionlp13cg-md==0.4.0
filelock==3.0.12
huggingface-hub==0.0.8
idna==2.10
isort==5.9.3
javabridge==1.0.19
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
liac-arff==2.5.0
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
murmurhash==1.0.5
nltk==3.6.2
nmslib==2.1.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==8.3.1
platformdirs==2.2.0
preshed==3.0.5
psutil==5.8.0
pybind11==2.6.1
pydantic==1.7.4
pylint==2.10.2
pyparsing==2.4.7
pysbd==0.3.4
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-weka-wrapper3==0.2.3
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
scispacy==0.4.0
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==3.0.0
spacy==3.0.6
spacy-alignments==0.8.3
spacy-legacy==3.0.6
spacy-transformers==1.0.3
srsly==2.4.1
tabulate==0.8.9
thinc==8.0.6
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
tqdm==4.61.1
transformers==4.6.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wasabi==0.8.2
wordcloud==1.8.1
wrapt==1.12.1
